// src/generated/prisma/models/Doctor.ts
export type DoctorWhereInput = {
  AND?: Prisma.DoctorWhereInput | Prisma.DoctorWhereInput[]
  OR?: Prisma.DoctorWhereInput[]
  NOT?: Prisma.DoctorWhereInput | Prisma.DoctorWhereInput[]
  doctorId?: Prisma.UuidFilter<"Doctor"> | string
  specialization?: Prisma.StringFilter<"Doctor"> | string
  appointments?: Prisma.AppointmentListRelationFilter
  doctorWeeklySlots?: Prisma.DoctorWeeklySlotListRelationFilter
  user?: Prisma.XOR<Prisma.UserScalarRelationFilter, Prisma.UserWhereInput>
}