// src/generated/prisma/sql/getWeeklySlots.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
import * as $runtime from "@prisma/client/runtime/client"

/**
 * @param uuid
 */
export const getWeeklySlots = $runtime.makeTypedQueryFactory("SELECT id,\nweekday,\nlower(time_range) AS \"startAt\",\nupper(time_range) AS \"endAt\"\nFROM doctor_weekly_slots\nWHERE doctor_id = $1\nORDER BY weekday ASC, \"startAt\" ASC;\n\n") as (uuid: string) => $runtime.TypedSql<getWeeklySlots.Parameters, getWeeklySlots.Result>

export namespace getWeeklySlots {
  export type Parameters = [uuid: string]
  export type Result = {
    id: string
    weekday: number
    startAt: Date | null
    endAt: Date | null
  }
}